Run over 2021 data on EOS. Move get chain function to tools.
